requirements.txt
alembic==1.3.1
aniso8601==6.0.0
attrs==19.3.0
awscli==1.18.4
Babel==2.8.0
botocore==1.15.4
cffi==1.14.0
Click==7.0
colorama==0.4.3
cryptography==2.8
docutils==0.15.2
ecdsa==0.15
Flask==1.1.1
Flask-CLI==0.4.0
Flask-Cors==3.0.7
Flask-Migrate==2.5.2
Flask-Moment==0.9.0
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.4.0
Flask-WTF==0.14.2
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.10.1
jmespath==0.9.4
jsonify==0.5
Mako==1.1.0
MarkupSafe==1.1.1
more-itertools==8.2.0
packaging==20.1
pluggy==0.13.1
psycopg2-binary==2.8.2
py==1.8.1
pyasn1==0.4.8
pycparser==2.19
PyJWT==1.7.1
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.6.0
python-dotenv==0.11.0
python-editor==1.0.4
python-jose==3.1.0
pytz==2019.1
PyYAML==5.2
rsa==3.4.2
s3transfer==0.3.3
six==1.12.0
SQLAlchemy==1.3.4
urllib3==1.25.8
virtualenv==16.7.9
wcwidth==0.1.8
Werkzeug==1.0.0
WTForms==2.2.1
